requirements.txt
tweepy
feedparser
python-dotenv
